backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-filter==24.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
dj-database-url==2.2.0
